requirements.txt
PyYAML==3.11
pytest